requirements.txt
streamlit
pandas
databricks-sql-connector
plotly
